requirements.txt
numpy-stl
scikit-image
numpy
matplotlib
scipy
solidpython
